requirements.txt
PyOpenGL
pyopengltk
kociemba
